+++ src/react/components/autocompleter/XUIAutocompleterSecondarySearch.d.ts
@@ -102,7 +102,7 @@ interface Props {
    * **Note:** *XUI design has also decided to keep a minimum width on the dropdown, so the dropdown
    * may not match the width of narrow triggers (setting this to `'min'` will not override this).*
    */
-  matchTriggerWidth?: boolean | true | false | 'min';
+  matchTriggerWidth?: true | false | 'min';
   /**
    * Callback for when the list closes.
    */