+++ src/react/components/autocompleter/XUIAutocompleterSecondarySearch.js
@@ -327,7 +327,7 @@ XUIAutocompleterSecondarySearch.propTypes = {
    * XUI design has also decided to keep a minimum width on the dropdown,
    * so dropdown may not match the width of narrow triggers (setting this to `'min'` will not override this).
    */
-  matchTriggerWidth: PropTypes.oneOfType([PropTypes.bool, PropTypes.oneOf([true, false, 'min'])]),
+  matchTriggerWidth: PropTypes.oneOf([true, false, 'min']),
 
   /** Callback for when the list closes */
   onClose: PropTypes.func,